\docType{data}
\name{craterA}
\alias{craterA}
\title{Bomb crater Point Pattern}
\format{An object of class \code{"ppp"} representing a point pattern of bomb craters. The Cartesian coordinates are in meters.
        See \code{\link[spatstat]{ppp.object}} for details of the format of a point pattern object.}
\usage{
data(craterA)
}
\description{
Bomb crater Point Pattern
}
\keyword{datasets}

