% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{craterB}
\alias{craterB}
\title{Bomb crater Point Pattern}
\format{An object of class \code{"ppp"} representing a point pattern of bomb craters. The Cartesian coordinates are in meters.
        See \code{\link[spatstat.geom]{ppp.object}} for details of the format of a point pattern object.}
\usage{
data(craterB)
}
\description{
Bomb crater Point Pattern
}
\keyword{datasets}
