% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dethrzbase.R
\name{det_alpha}
\alias{det_alpha}
\title{calculation of alpha (failure probability), when having the threshold c}
\usage{
det_alpha(intens, threshold, nxprob = 0.1)
}
\arguments{
\item{intens}{estimated intensity of the observed process (object of class "im", see \code{\link[spatstat.core]{density.ppp}})}

\item{threshold}{threshold c: The high-risk zone is the field in which the estimated intensity 
exceeds this value.}

\item{nxprob}{probability of having unobserved events}
}
\value{
value of alpha
}
\description{
This function is used for the intensity-based method. It determines the
probability to have at least one unobserved event outside the high-risk zone. 
A Poisson distribution is used for the number of unobserved events in a certain area or field.
Used in functions det_threshold, det_thresholdfromarea.
}
