% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_hrz_eval_ar.R
\name{det_alpha_eval_ar}
\alias{det_alpha_eval_ar}
\title{Determination of failure probability within evaluation area}
\usage{
det_alpha_eval_ar(intens, eval_ar, threshold, nxprob = 0.1)
}
\arguments{
\item{intens}{estimated intensity}

\item{eval_ar}{evaluation area}

\item{threshold}{given threshold}

\item{nxprob}{constant probability of non-explosion}
}
\description{
Determination of failure probability within evaluation area
}
