% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dethrz.R
\name{det_hrz}
\alias{det_hrz}
\alias{highriskzone.object}
\alias{highriskzone}
\title{Determination of the high-risk zone.}
\usage{
det_hrz(ppdata, type, criterion, cutoff, distancemap = NULL, intens = NULL,
  nxprob = 0.1, covmatrix = NULL)
}
\arguments{
\item{ppdata}{Observed spatial point process of class ppp.}

\item{type}{Method to use, can be one of \code{"dist"} (method of fixed radius or quantile-based method), or
\code{"intens"} (intensity-based method)}

\item{criterion}{criterion to limit the high-risk zone, can be one of 
\code{"area"} (giving size of hrz), \code{"indirect"} (giving quantile/alpha depending on type),
or \code{"direct"} (giving radius/threshold c depending on type)}

\item{cutoff}{Value of criterion (area, radius, quantile, alpha or threshold). 
Depending on criterion and type: 
If criterion = "direct" and type = "intens", cutoff is the maximum intensity of unexploded bombs
outside the risk zone. If type = "dist" instead, cutoff is the radius of the circle around each exploded bomb.
"If criterion = "indirect", cutoff is the quantile for the quantile-based method and the failure 
probability alpha for the intensity-base method. If criterion = "area", cutoff is the area the high-risk zone should
have.}

\item{distancemap}{(optional) distance map: distance of every pixel to the nearest observation 
of the point pattern; only needed for \code{type="dist"}. If not given, 
it will be computed by \code{\link[spatstat.geom]{distmap}}.}

\item{intens}{(optional) estimated intensity of the observed process (object of class "im"), 
only needed for type="intens". If not given,
it will be estimated using \code{\link[spatstat.core]{density.ppp}}.}

\item{nxprob}{Probability of having unobserved events.
Default value is 0.1.}

\item{covmatrix}{(optional) Covariance matrix of the kernel of a normal distribution, only needed for 
\code{type="intens"} if no intensity is given. If not given, it will be estimated
using \code{\link[ks]{Hscv}}.}
}
\value{
An object of class "\code{highriskzone}", which is a list of
   \item{ typehrz, criterion, cutoff, nxprob }{ see arguments}
   \item{ zone }{ Determined high-risk zone: Object of class "owin" based on a binary mask. 
                  See \code{\link[spatstat.geom]{owin}}. }
   \item{ threshold }{ determined threshold. If type = "dist" and criterion = "direct" it is the specified radius.
   If criterion = "indirect" or "area" the determined radius used to construct a risk zone fulfilling the specified criterion 
   and cutoff. If type = "dist" it is the specified or calculated threshold c, the maximum intensitiy of unexploded bombs 
   outside the risk zone.}
   \item{ calccutoff }{ determined cutoff-value. For type="dist" and criterion="area", this is the
quantile of the nearest-neighbour distance. For type="intens" and criterion="area" or "direct", it is the failure
probability alpha. For all other criterions it is NA.}
   \item{ covmatrix }{ If not given (and \code{type="intens"}), it is estimated. See \code{\link[ks]{Hscv}}.}
}
\description{
\code{det_hrz} determines the high-risk zone through the method of fixed radius 
(type = "dist" and criterion = "direct"), the quantile-based method (type = "dist" and 
criterion = "area"/"indirect") and the intensity-based method (type =  "intens").
}
\details{
There are different methods implemented to determine a high-risk zone.
\describe{
\item{ Method of fixed radius }{
       In this method, the high-risk zone is determined by drawing a circle around each
       observed event with a fixed radius. This method will be used when \code{type = "dist"}
       and \code{criterion = "direct"}. \code{cutoff} then is the radius.
       }
\item{ Quantile-based method }{
       This method is a development of the above. Here the radius is not fixed. It uses
       the distance of every observed event to the nearest other event, which is calculated by the
       nearest-neighbour distance. The radius is assessed by the p-quantile of the empirical
       distribution function of the nearest-neighbour distance. This method will be used when
       \code{type = "dist"} and \code{criterion = "indirect"} or \code{"area"}. If 
       \code{criterion = "indirect"}, then \code{cutoff} is the quantile that should be used.
       If \code{criterion = "area"} then \code{cutoff} is the area that the  high-risk zone 
       has to have at the end and from that the quantile/the radii are determined. When the
       calculation is done via the area, it can not really be classified to the quantile-based
       method. It is rather a third "distance-based" method.
       }
\item{ Intensity-based method }{
       The first step of this method is to estimate the intensity of the observed events.
       Based on the estimated intensity and the specified probability of unobserved bombs \code{nxprob} 
       it is possible to estimate the intensity of unobserved/unexploded bombs.
       The high-risk zone is then the area in which the estimated intensity of unexploded bombs exceeds a 
       certain  value. This value is called threshold c.
       The method will be used when \code{type = "intens"}. There are three different ways to
       construct a high-risk zone: 
       \enumerate{
          \item Fixing the threshold c: \code{criterion = "direct"} 
          \item Fixing the area of the high-risk zone: \code{criterion = "area"}
          \item Fixing the failure probability alpha, which is the probability of having 
                unobserved events outside the high-risk zone: \code{criterion = "indirect"}
                Here, the point process is assumed to be an inhomogeneous Poisson process.
         }
       For further information see Mahling et al. (2013) (References).
       }
}

If there are restriction areas in the observation window, use \code{\link[highriskzone]{det_hrz_restr}}
instead. For estimation of intensity based highrikszones with a bigger observation area than area of interest
(evaluation area) use \code{\link[highriskzone]{det_hrz_eval_ar}}.
}
\examples{
 data(craterA)
## change npixel to 1000 to obtain nicer plots
spatstat.geom::spatstat.options(npixel=100)
## type: dist
hrzd1 <- det_hrz(craterA, type = "dist", criterion = "area", cutoff = 1000000, nxprob = 0.1)
hrzd2 <- det_hrz(craterA, type = "dist", criterion = "indirect", cutoff = 0.9, nxprob = 0.1)
hrzd3 <- det_hrz(craterA, type = "dist", criterion = "direct", cutoff = 100, nxprob = 0.1)

op <- par(mfrow = c(2, 2))
plot(craterA)
plot(hrzd1, zonecol = 2, win = craterA$window, plotwindow = TRUE)
plot(hrzd2, zonecol = 3,  win = craterA$window, plotwindow = TRUE)
plot(hrzd3, zonecol = 4,  win = craterA$window, plotwindow = TRUE)
par(op)

\dontrun{
# or first calculate the distancemap and use it:
distm <- distmap(craterA)
hrzd <- det_hrz(craterA, type = "dist", criterion = "direct", cutoff = 100,
                distancemap = distm, nxprob = 0.1)
}                
## type: intens 
# reduce number of observations for faster computation
thin.craterA <- craterA[1:10]
hrzi1 <- det_hrz(thin.craterA, type = "intens", criterion = "area", cutoff = 100000, nxprob = 0.1)
plot(hrzi1)
plot(thin.craterA, add = TRUE)
plot(thin.craterA$window, add = TRUE)
\dontrun{
hrzi2 <- det_hrz(craterA, type = "intens", criterion = "indirect", cutoff = 0.1, nxprob = 0.1)
hrzi3 <- det_hrz(craterA, type = "intens", criterion = "direct", cutoff = 0.0001, nxprob = 0.1)
plot(hrzi2)
plot(hrzi3)
}
                 
## More detailed examples on http://highriskzone.r-forge.r-project.org/
}
\references{
Monia Mahling, Michael \enc{Hoehle}{Hoehle} & Helmut \enc{Kuechenhoff}{Kuechenhoff} (2013),
\emph{Determining high-risk zones for unexploded World War II bombs by using point process methodology.}
Journal of the Royal Statistical Society, Series C 62(2), 181-199.

Monia Mahling (2013),
\emph{Determining high-risk zones by using spatial point process methodology.}
Ph.D. thesis, Cuvillier Verlag \enc{Goettingen}{Goettingen},
available online: http://edoc.ub.uni-muenchen.de/15886/
}
\seealso{
\code{\link[spatstat.geom]{distmap}}, \code{\link[spatstat.geom]{eval.im}}, \code{\link[spatstat.geom]{owin}},
               \code{\link{eval_method}}, \code{\link[highriskzone]{det_hrz_restr}}
}
