% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dethrzbase.R
\name{det_nnarea}
\alias{det_nnarea}
\title{Determination of the area of a high-risk zone using the nearest-neighbour distance.}
\usage{
det_nnarea(cutoffval, distancemap, win)
}
\arguments{
\item{cutoffval}{distance used as radius of the discs}

\item{distancemap}{distance map (object of class "im", see \code{\link[spatstat.geom]{distmap}}): 
distance of every location in the observation window to the nearest event}

\item{win}{observation window of class owin}
}
\value{
A numerical value giving the area of the window.
}
\description{
Used in function det_radius.
}
\seealso{
\code{\link[spatstat.geom]{eval.im}}, \code{\link[spatstat.geom]{owin}}, \code{\link[spatstat.geom]{area.owin}}
}
