% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationbase.R
\name{det_nsintens}
\alias{det_nsintens}
\title{Determination of the intensity for the Neyman Scott simulation.}
\usage{
det_nsintens(ppdata, radius)
}
\arguments{
\item{ppdata}{observed point pattern whose estimated intensity (adjusted for
thinning and divided by "clustering") is used for simulating the
parent process}

\item{radius}{radius of the circles around the parent points in which the cluster
points are located}
}
\value{
A pixel image (object of class "im"). See \code{\link[spatstat.core]{density.ppp}}.
}
\description{
Used in function sim_nsppp.
}
\seealso{
\code{\link[spatstat.core]{density.ppp}}, \code{\link[spatstat.geom]{boundingbox}}, 
         \code{\link[spatstat.geom]{owin}}, \code{\link[ks]{Hscv}}
}
