% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detnsintens_restr.R
\name{det_nsintens_restr}
\alias{det_nsintens_restr}
\title{Determination of the intensity for the Neyman-Scott simulation.}
\usage{
det_nsintens_restr(ppdata, radius, weights)
}
\arguments{
\item{ppdata}{observed point pattern whose estimated intensity (adjusted for
thinning and divided by "clustering") is used for simulating the
parent process}

\item{radius}{radius of the circles around the parent points in which the cluster
points are located}

\item{weights}{Vector of observation probabilities associated with the observations contained in \code{ppdata}.}
}
\value{
A pixel image (object of class "im"). See \code{\link[spatstat.core]{density.ppp}}.
}
\description{
Used in function bootcor_restr.
}
\seealso{
\code{\link[spatstat.core]{density.ppp}}, \code{\link[spatstat.geom]{boundingbox}},
         \code{\link[spatstat.geom]{owin}}, \code{\link[ks]{Hscv}}
}
