% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dethrzbase.R
\name{det_radius}
\alias{det_radius}
\title{Determination of the nearest-neighbour distance which results in a high-risk zone with desired area}
\usage{
det_radius(ppdata, distancemap, areahrz, win)
}
\arguments{
\item{ppdata}{observed spatial point pattern of class ppp.}

\item{distancemap}{distance map (object of class "im", see \code{\link[spatstat.geom]{distmap}}): 
distance of every location in the observation window to the nearest event}

\item{areahrz}{given area of the high-risk zone}

\item{win}{observation window of class owin}
}
\value{
A list of 
    \item{ cutoffdist }{ quantile of the nearest-neighbour distance }
    \item{ thresh }{ distance }
}
\description{
Used in function det_hrz.
}
\seealso{
\code{\link{det_nnarea}}, \code{\link[stats]{quantile}}, \code{\link[stats]{uniroot}}
}
