% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dethrzbase.R
\name{det_threshold}
\alias{det_threshold}
\title{Calculation of the threshold c, when having failure probability alpha.}
\usage{
det_threshold(intens, alpha = 1e-05, nxprob = 0.1)
}
\arguments{
\item{intens}{estimated intensity of the observed process (object of class "im", see \code{\link[spatstat.core]{density.ppp}})}

\item{alpha}{failure probability: probability to have at least one unobserved event
outside the high-risk zone}

\item{nxprob}{probability of having unobserved events}
}
\value{
value of the threshold c
}
\description{
The high-risk zone is the field in which the estimated intensity 
exceeds the threshold c, which is determined here, having the failure probability
alpha.
This function is for the intensity-based method. Used in function det_hrz.
}
\seealso{
\code{\link{det_alpha}}, \code{\link[stats]{uniroot}}
}
