% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_hrz_eval_ar.R
\name{det_threshold_eval_ar}
\alias{det_threshold_eval_ar}
\title{Determination of necessary threshold to keep alpha in evaluation area}
\usage{
det_threshold_eval_ar(intens, eval_ar, alpha = 1e-05, nxprob = 0.1)
}
\arguments{
\item{intens}{estimated intensity}

\item{eval_ar}{evaluation area}

\item{alpha}{desired failure probability in eval area}

\item{nxprob}{constant probability of non-explosion}
}
\description{
Determination of necessary threshold to keep alpha in evaluation area
}
