% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generichrz.R
\name{print.highriskzone}
\alias{print.highriskzone}
\title{Print Brief Details of a high-risk zone}
\usage{
\method{print}{highriskzone}(x, ...)
}
\arguments{
\item{x}{high-risk zone (object of class "\code{highriskzone}")}

\item{...}{ignored}
}
\description{
Prints a very brief description of a high-risk zone.
}
\details{
A very brief description of the highriskzone x is printed.
This is a method for the generic function \code{\link[base]{print}}.
}
\seealso{
\code{\link[base]{print}}, \code{\link{summary.highriskzone}}
}
