% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationbase.R
\name{sim_intens}
\alias{sim_intens}
\title{Simulation on given intensity}
\usage{
sim_intens(ppdata, intensSim, nxprob)
}
\arguments{
\item{ppdata}{Observed spatial point process of class ppp}

\item{intensSim}{Intensity to use for the simulation}

\item{nxprob}{Probability of having unobserved events}
}
\value{
A list of of observed and unobserved point patterns (see \code{\link{thin}})
}
\description{
Generation of a random point pattern using the inhomogeneous Poisson process (if lambda is not constant)
and thinning of this data, to obtain "observed" and "unobserved" events.
}
\seealso{
\code{\link{thin}}, \code{\link[spatstat]{rpoispp}}
}
