% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericbootcorr.R
\name{summary.bootcorr}
\alias{summary.bootcorr}
\title{Summary of a the bootstrap correction for a high-risk zone}
\usage{
\method{summary}{bootcorr}(object, ...)
}
\arguments{
\item{object}{bootstrap correction for a high-risk zone (object of class "\code{bootcorr}")}

\item{...}{ignored}
}
\description{
Prints a useful summary of the bootstrap correction for a high-risk zone.
}
\details{
A useful summary of the bootstrap correction x for a high-risk zone is printed.
This is a method for the generic function \code{\link[base]{summary}}.
}
\seealso{
\code{\link[base]{summary}}, \code{\link{print.bootcorr}}, \code{\link{plot.bootcorr}}
}
