% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generichrz.R
\name{summary.highriskzone}
\alias{summary.highriskzone}
\title{Summary of a high-risk zone}
\usage{
\method{summary}{highriskzone}(object, ...)
}
\arguments{
\item{object}{high-risk zone (object of class "\code{highriskzone}")}

\item{...}{ignored}
}
\description{
Prints a useful summary of a high-risk zone.
}
\details{
A useful description of the highriskzone object is printed.
This is a method for the generic function \code{\link[base]{summary}}.
}
\seealso{
\code{\link[base]{summary}}, \code{\link{print.highriskzone}}
}
