% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericevalhrz.R
\name{summary.hrzeval}
\alias{summary.hrzeval}
\title{Summary of a the evaluation of a high-risk zone}
\usage{
\method{summary}{hrzeval}(object, ...)
}
\arguments{
\item{object}{evaluation of a high-risk zone (object of class "\code{hrzeval}")}

\item{...}{ignored}
}
\description{
Prints a useful summary of the evaluation of a high-risk zone.
}
\details{
A useful description of the hrzeval object is printed.
This is a method for the generic function \code{\link[base]{summary}}.
}
\seealso{
\code{\link[base]{summary}}, \code{\link{print.hrzeval}}
}
