% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationbase.R
\name{thin}
\alias{thin}
\title{Thinning of the observations (for evaluating the method)}
\usage{
thin(full, nxprob)
}
\arguments{
\item{full}{all observations of the point pattern}

\item{nxprob}{probability of having unobserved events}
}
\value{
A list of observed and unobserved point patterns. Both of class ppp.
}
\description{
The thinning is done by drawing independently from a Bernoulli distribution. This function is needed for functions eval_method, sim_clintens, sim_intens
}
\examples{
 data(craterB)
 thdata <- thin(craterB, nxprob=0.1)
 thdata
 plot(thdata$observed); points(thdata$unobserved, col=4)
}
\seealso{
\code{\link[stats]{rbinom}}, \code{\link[spatstat.geom]{ppp}}
}
