% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highriskzone-package.R
\docType{package}
\encoding{latin1}
\name{highriskzone-package}
\alias{highriskzone-package}
\alias{package-highriskzone}
\title{Determining high-risk zones by using spatial point process methodology}
\description{
The package highriskzone provides tools to determine and evaluate high-risk zones of
unobserved events by using point process methodology.
}
\references{
Monia Mahling, Michael \enc{Hoehle}{Hoehle} & Helmut \enc{Kuechenhoff}{Kuechenhoff} (2013),
\emph{Determining high-risk zones for unexploded World War II bombs by using point process methodology.}
Journal of the Royal Statistical Society, Series C 62(2), 181-199. \cr

Monia Mahling (2013),
\emph{Determining high-risk zones by using spatial point process methodology.}
Ph.D. thesis, Cuvillier Verlag \enc{Goettingen}{Goettingen},
available online: http://edoc.ub.uni-muenchen.de/15886/\cr

Heidi Seibold (2012), \emph{Determining high risk zones using point process methodology 
- Realization by building an R package.} Bachelor Thesis, Ludwig Maximilian University of Munich.
}
\seealso{
\code{\link[spatstat]{spatstat-package}}
}
\author{
Heidi Seibold \email{Heidi.Seibold@campus.lmu.de},
Monia Mahling \email{monia.mahling@stat.uni-muenchen.de}
Sebastian Linne \email{Sebastian.Linne@campus.lmu.de}
Felix Guenther \email{felix.guenther@stat.uni-muenchen.de}
Maintainer: Felix Guenther \email{felix.guenther@stat.uni-muenchen.de}
}
\keyword{package}
